feature: Use the newer NetworkCallback API on versions of Android which support it
